Updated initial state location in reducer to accomadate testing
Local storage does not exist in the mocha/jsdom implementation so it must be
mocked as a constant in the test_helper. Initial state for reducers need to
be moved to functions so it is not executed on test run - if this happens the
localStorage object is not instantiated yet.